feat: support json rpc proxy
